プルリクで指摘されたところ修正しました

+++ db/migrate/20190523001841_create_influencers.rb
@@ -2,7 +2,7 @@ class CreateInfluencers < ActiveRecord::Migration[5.2]
   def change
     create_table :influencers do |t|
       t.string :name, null: false, default: ''
-      t.integer :ito_id, null: false
+      t.bigint :ito_id, null: false
       t.timestamps
     end
   end
