+++ db/schema.rb
@@ -22,7 +22,7 @@ ActiveRecord::Schema.define(version: 2019_05_23_001841) do
 
   create_table "influencers", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "name", default: "", null: false
-    t.integer "ito_id", null: false
+    t.bigint "ito_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
